refactor(lynx-react): move React hooks to src/hooks/
Relocate useControllableState and usePressTap (plus their tests) from
src/utils/ to src/hooks/ so that hook files are no longer mixed with pure
utilities. Update the three internal import paths (index.ts, Switch.tsx,
ActionButton.tsx) to the new location.

Public API (re-exported symbols and types) is unchanged — only the internal
relative path changed.
