+++ packages/lynx-react/src/components/Switch/Switch.tsx
@@ -8,8 +8,8 @@ import type { SwitchmarkVariantProps } from "@seed-design/lynx-css/recipes/switc
 
 import { createCompoundContext } from "../../utils/create-compound-context";
 import { splitMultipleVariantsProps } from "../../utils/split-multiple-variants-props";
-import { useControllableState } from "../../utils/use-controllable-state";
-import { usePressTap } from "../../utils/use-press-tap";
+import { useControllableState } from "../../hooks/use-controllable-state";
+import { usePressTap } from "../../hooks/use-press-tap";
 
 /**
  * @platform Lynx
